Added Security and enhanced OrderController to be perfect for ThymeLeaf.

+++ src/main/java/com/Group6/WebAppDevGroupProject/Models/MenuItem.java
@@ -9,21 +9,21 @@ import jakarta.persistence.Id;
 public class MenuItem {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int item_id;
+    private long item_id;
     private String name;
     private String description;
     private float price;
     private int stock;
 
     // -- Basic Getters --
-    public int getItem_id() { return item_id; }
+    public long getItem_id() { return item_id; }
     public String getName() { return name; }
     public String getDescription() { return description; }
     public float getPrice() { return price; }
     public int getStock() { return stock; }
 
     // -- Basic Setters --
-    public void setItem_id(int item_id) { this.item_id = item_id; }
+    public void setItem_id(long item_id) { this.item_id = item_id; }
     public void setName(String name) { this.name = name; }
     public void setDescription(String description) { this.description = description; }
     public void setPrice(float price) { this.price = price; }
