+++ src/main/java/com/Group6/WebAppDevGroupProject/Models/User.java
@@ -9,21 +9,21 @@ import jakarta.persistence.Id;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int user_id;
+    private long user_id;
     private String username;
     private String email;
     private String password;
     private String role;
 
     // -- Basic Getters --
-    public int getUser_id() { return user_id; }
+    public long getUser_id() { return user_id; }
     public String getUsername() { return username; }
     public String getEmail() { return email; }
     public String getPassword() { return password; }
     public String getRole() { return role; }
 
     // -- Basic Setters --
-    public void setUser_id(int user_id) { this.user_id = user_id; }
+    public void setUser_id(long user_id) { this.user_id = user_id; }
     public void setUsername(String username) { this.username = username; }
     public void setEmail(String email) { this.email = email; }
     public void setPassword(String password) { this.password = password; }
